move log begin logic
