+++ crates/log/Cargo.toml
@@ -20,3 +20,5 @@ bevy.workspace = true
 flate2.workspace = true
 tracing-appender.workspace = true
 tracing-subscriber.workspace = true
+chrono.workspace = true
+tracing.workspace = true
