+++ src/main.rs
@@ -26,52 +26,26 @@ use de_signs::SignsPluginGroup;
 use de_spawner::SpawnerPluginGroup;
 use de_terrain::TerrainPluginGroup;
 use tracing::{span, Level};
-use tracing_subscriber::layer::SubscriberExt;
-use tracing_subscriber::{fmt, EnvFilter};
 
 const CARGO_PKG_VERSION: &str = env!("CARGO_PKG_VERSION");
 const GIT_SHA: &str = env!("GIT_SHA");
 
 fn main() {
-    // TODO: move log init to de_log crate
-    // for file name
-    let dt = chrono::Local::now();
-
-    let file_name = dt.format("%Y-%m-%d %H:%M:%S.log").to_string();
-    let file_appender = tracing_appender::rolling::never("logs", &*file_name);
-
-    let (non_blocking_log_writer, _guard) = tracing_appender::non_blocking(file_appender);
-
-    let collector = tracing_subscriber::registry()
-        .with(
-            EnvFilter::from_default_env()
-                .add_directive(Level::TRACE.into())
-                .add_directive("wgpu_core=warn".parse().unwrap())
-                .add_directive("bevy_ecs=info".parse().unwrap())
-                .add_directive("async_io=info".parse().unwrap())
-                .add_directive("polling=debug".parse().unwrap())
-                .add_directive("wgpu_hal=info".parse().unwrap())
-                .add_directive("naga=info".parse().unwrap()),
-        )
-        .with(fmt::layer().with_writer(std::io::stdout))
-        .with(fmt::layer().with_writer(non_blocking_log_writer));
-    tracing::subscriber::set_global_default(collector).expect("Unable to set a global collector");
-
+    let mut app = App::new();
+    // we want logging as early as possible
+    app.add_plugins(LogPluginGroup);
+    
     info!(
         "Starting Digital Extinction {{ \"Version\": \"{}\", \"GitSha\": \"{}\" }}",
         CARGO_PKG_VERSION, GIT_SHA
     );
 
-    // so we can use it after span is closed
-    let mut app: App;
 
     {
         let span = span!(Level::TRACE, "Startup");
         let _enter = span.enter();
 
-        app = App::new();
         app.insert_resource(Msaa::Sample4)
-            .add_plugins(LogPluginGroup)
             .add_plugins(
                 DefaultPlugins
                     .set(WindowPlugin {
